+++ packages/scenes/src/core/SceneObjectBase.tsx
@@ -1,5 +1,5 @@
 import { useEffect } from 'react';
-import { Observer, Subject, Subscription, SubscriptionLike, Unsubscribable } from 'rxjs';
+import { Subscription, Unsubscribable } from 'rxjs';
 import { v4 as uuidv4 } from 'uuid';
 
 import { BusEvent, BusEventHandler, BusEventType, EventBusSrv } from '@grafana/data';
@@ -8,7 +8,9 @@ import {
   SceneComponent,
   SceneObjectState,
   SceneObjectUrlSyncHandler,
-  ExternalActivationOrDeactivationHandler,
+  SceneStateChangedHandler,
+  SceneActivationHandler,
+  SceneDeactivationHandler,
 } from './types';
 import { useForceUpdate } from '@grafana/ui';
 
@@ -21,11 +23,10 @@ export abstract class SceneObjectBase<TState extends SceneObjectState = SceneObj
   implements SceneObject<TState>
 {
   private _isActive = false;
-  private _subject = new Subject<TState>();
   private _state: TState;
   private _events = new EventBusSrv();
-  private _activationHandlers: ExternalActivationOrDeactivationHandler[] = [];
-  private _deactivationHandlers: ExternalActivationOrDeactivationHandler[] = [];
+  private _activationHandlers: SceneActivationHandler[] = [];
+  private _deactivationHandlers: SceneDeactivationHandler[] = [];
 
   protected _parent?: SceneObject;
   protected _subs = new Subscription();
@@ -39,7 +40,6 @@ export abstract class SceneObjectBase<TState extends SceneObjectState = SceneObj
     }
 
     this._state = Object.freeze(state);
-    this._subject.next(state);
     this.setParent();
   }
 
@@ -90,8 +90,12 @@ export abstract class SceneObjectBase<TState extends SceneObjectState = SceneObj
   /**
    * Subscribe to the scene state subject
    **/
-  public subscribeToState(observerOrNext?: Partial<Observer<TState>>): SubscriptionLike {
-    return this._subject.subscribe(observerOrNext);
+  public subscribeToState(handler: SceneStateChangedHandler<TState>): Unsubscribable {
+    return this._events.subscribe(SceneObjectStateChangedEvent, (event) => {
+      if (event.payload.changedObject === this) {
+        handler(event.payload.newState as TState, event.payload.prevState as TState);
+      }
+    });
   }
 
   /**
@@ -111,7 +115,6 @@ export abstract class SceneObjectBase<TState extends SceneObjectState = SceneObj
     this._state = Object.freeze(newState);
 
     this.setParent();
-    this._subject.next(newState);
 
     // Bubble state change event. This is event is subscribed to by UrlSyncManager and UndoManager
     this.publishEvent(
@@ -194,9 +197,6 @@ export abstract class SceneObjectBase<TState extends SceneObjectState = SceneObj
     this._events.removeAllListeners();
     this._subs.unsubscribe();
     this._subs = new Subscription();
-
-    this._subject.complete();
-    this._subject = new Subject<TState>();
   }
 
   /**
@@ -223,7 +223,7 @@ export abstract class SceneObjectBase<TState extends SceneObjectState = SceneObj
    * Allows external code to register code that is executed on activate and deactivate. This allow you
    * to wire up scene objects that need to respond to state changes in other objects from the outside.
    **/
-  public registerActivationHandler(handler: ExternalActivationOrDeactivationHandler) {
+  public registerActivationHandler(handler: SceneActivationHandler) {
     this._activationHandlers.push(handler);
   }
 }
@@ -236,7 +236,7 @@ function useSceneObjectState<TState extends SceneObjectState>(model: SceneObject
   const forceUpdate = useForceUpdate();
 
   useEffect(() => {
-    const s = model.subscribeToState({ next: forceUpdate });
+    const s = model.subscribeToState(forceUpdate);
     return () => s.unsubscribe();
   }, [model, forceUpdate]);